Projects/App.js: move logic to NewIssueCuration
This allows moving opening of the panel to usePanelManagement hook
